--- data/src/main/java/com/madrid/data/dataSource/local/SearchLocalSource.kt
@@ -1,5 +1,0 @@
-package com.madrid.data.dataSource.local
-
-interface SearchLocalSource {
-
-}
