+++ data/src/main/java/com/madrid/data/dataSource/local/dao/MovieDao.kt
@@ -1,4 +1,4 @@
-package com.example.data.dataSource.local.data.dao
+package com.madrid.data.dataSource.local.dao
 
 import androidx.room.Dao
 import androidx.room.Delete
@@ -6,7 +6,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import com.example.data.dataSource.local.data.entity.MovieEntity
+import com.madrid.data.dataSource.local.entity.MovieEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
