+++ data/src/main/java/com/madrid/data/dataSource/local/dao/SeriesDao.kt
@@ -1,12 +1,12 @@
-package com.example.data.dataSource.local.data.dao
+package com.madrid.data.dataSource.local.dao
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.example.data.dataSource.local.data.entity.MovieEntity
-import com.example.data.dataSource.local.data.entity.SeriesEntity
+import com.madrid.data.dataSource.local.entity.MovieEntity
+import com.madrid.data.dataSource.local.entity.SeriesEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
