+++ data/src/main/java/com/madrid/data/repositories/SearchLocalSource.kt
@@ -1,6 +1,6 @@
-package com.example.data.repositories
+package com.madrid.data.repositories
 
-import com.example.data.dataSource.local.data.entity.MovieEntity
+import com.madrid.data.dataSource.local.entity.MovieEntity
 import kotlinx.coroutines.flow.Flow
 
 interface SearchLocalSource {
