+++ data/src/main/java/com/madrid/data/repositories/mappers/SeriesMappers.kt
@@ -1,33 +1,32 @@
-package com.example.data.repositories.mappers
-
-import com.example.data.dataSource.local.data.entity.MovieEntity
-import com.example.domain.entity.Movie
+package com.madrid.data.repositories.mappers
 
 
-fun Movie.toMovieEntity(): MovieEntity {
-    return MovieEntity(
+import com.madrid.data.dataSource.local.entity.SeriesEntity
+import com.example.domain.entity.Series
+
+fun Series.toSeriesEntity(): SeriesEntity {
+    return SeriesEntity(
         id = this.id,
         title = this.title,
         imageUrl = this.imageUrl,
         rate = this.rate,
         yearOfRelease = this.yearOfRelease,
-        movieDuration = this.movieDuration,
         description = this.description,
     )
 }
 
-fun MovieEntity.toMovie(): Movie {
-    return Movie(
+fun SeriesEntity.toSeries(): Series {
+    return Series(
         id = this.id,
         title = this.title,
         imageUrl = this.imageUrl,
         rate = this.rate,
         yearOfRelease = this.yearOfRelease,
-        movieDuration = this.movieDuration,
         description = this.description,
         genre = TODO(),
         topCast = TODO(),
         reviews = TODO(),
+        seasons = TODO(),
     )
 }
